address test failures
This commit addresses test failures by declaring explicit Decode/EncodeBytes
types. The logic around implicit conversion was too tricky to get right,
considering how many we have defined for `Event` already.
