+++ lib/vector-core/src/buffers/mod.rs
@@ -2,20 +2,16 @@ mod acker;
 #[cfg(feature = "disk-buffer")]
 pub mod disk;
 
+use crate::bytes::{DecodeBytes, EncodeBytes};
 pub use acker::Acker;
-use bytes::Bytes;
 use futures::{channel::mpsc, Sink, SinkExt, Stream};
 use pin_project::pin_project;
 use serde::{Deserialize, Serialize};
-use std::fmt::Debug;
+use std::fmt::{Debug, Display};
 #[cfg(feature = "disk-buffer")]
 use std::path::Path;
 use std::pin::Pin;
 use std::task::{Context, Poll};
-use std::{
-    convert::{TryFrom, TryInto},
-    fmt::Display,
-};
 
 // NOTE unfortunately because we can't edit out a lifetime based on a feature
 // flag we need two copies of `Variant` else the liftime being unused when
@@ -62,9 +58,9 @@ pub fn build<'a, T>(
     String,
 >
 where
-    T: 'a + Send + Sync + Unpin + Clone + TryInto<Bytes> + TryFrom<Bytes>,
-    <T as TryInto<bytes::Bytes>>::Error: Debug,
-    <T as TryFrom<bytes::Bytes>>::Error: Debug + Display,
+    T: 'a + Send + Sync + Unpin + Clone + EncodeBytes<T> + DecodeBytes<T>,
+    <T as EncodeBytes<T>>::Error: Debug,
+    <T as DecodeBytes<T>>::Error: Debug + Display,
 {
     match variant {
         #[cfg(feature = "disk-buffer")]
@@ -114,9 +110,9 @@ impl Default for WhenFull {
 #[derive(Clone)]
 pub enum BufferInputCloner<T>
 where
-    T: Send + Sync + Unpin + Clone + TryInto<Bytes> + TryFrom<Bytes>,
-    <T as TryInto<bytes::Bytes>>::Error: Debug,
-    <T as TryFrom<bytes::Bytes>>::Error: Debug,
+    T: Send + Sync + Unpin + Clone + EncodeBytes<T> + DecodeBytes<T>,
+    <T as EncodeBytes<T>>::Error: Debug,
+    <T as DecodeBytes<T>>::Error: Debug,
 {
     Memory(mpsc::Sender<T>, WhenFull),
     #[cfg(feature = "disk-buffer")]
@@ -125,9 +121,9 @@ where
 
 impl<'a, T> BufferInputCloner<T>
 where
-    T: 'a + Send + Sync + Unpin + Clone + TryInto<Bytes> + TryFrom<Bytes>,
-    <T as TryInto<bytes::Bytes>>::Error: Debug,
-    <T as TryFrom<bytes::Bytes>>::Error: Debug,
+    T: 'a + Send + Sync + Unpin + Clone + EncodeBytes<T> + DecodeBytes<T>,
+    <T as EncodeBytes<T>>::Error: Debug,
+    <T as DecodeBytes<T>>::Error: Debug + Display,
 {
     pub fn get(&self) -> Box<dyn Sink<T, Error = ()> + 'a + Send> {
         match self {
