+++ lib/vector-core/src/event/test/common.rs
@@ -1,0 +1,1 @@
+// common
