+++ lib/vector-core/src/lib.rs
@@ -24,6 +24,7 @@
 #[cfg(feature = "api")]
 pub mod api;
 pub mod buffers;
+pub mod bytes;
 pub mod config;
 pub mod event;
 pub mod mapping;
